SNNCode/requirements.txt
IPython
snntorch
nir
nirtorch
pandas
tqdm
wandb
matplotlib
torchinfo
scikit-learn

